Fix static analysis issues
:Release Notes:
Fix static analysis issues

:Detailed Notes:
Fix static analysis issues

:Testing Performed:
Unit Test Passed

:QA Notes:
NA

:Issues Addressed:
[PLAT-141842] Fix the static analays issue
[CID-8989067] MISRA C++-2008 Expressions

+++ src/ResBundle.cpp
@@ -248,7 +248,7 @@ const string& ResBundle::getPseudoString(const string& key, const string& source
 			const string DIGITS = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 			long total = 1;
 			for (int i = 0; i < source.length(); i++) {
-				total += source.at(i);
+				total += static_cast<int>(source.at(i));
 				total *= 157291;
 				total %= 238321;
 			}
